# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Styr API
  description: >
    Styr's REST and SSE API: OIDC login, Claude Code sessions, approvals,
    workspaces, profiles and server administration.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: /
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/v1/templates:
    get:
      summary: List templates
      description: Every template visible to the current user (owned, shared, or — for an admin — all).
      operationId: listTemplates
      tags: [templates]
      responses:
        "200":
          description: Every visible template
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      summary: Create a template
      description: >
        shared marks the template as owned by nobody rather than the
        caller; only an admin may set it.
      operationId: createTemplate
      tags: [templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/templates/{id}:
    get:
      summary: Get a template
      operationId: getTemplate
      tags: [templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update a template
      operationId: patchTemplate
      tags: [templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "200":
          description: The updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a template
      operationId: deleteTemplate
      tags: [templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/templates/{id}/render:
    post:
      summary: Dry-run render a template
      description: >
        Renders a title and prompt against payload without starting a
        session. kind selects how payload is normalized into template
        variables ("generic", "grafana" or "github"); it defaults to
        "grafana" when omitted. When title_template and/or prompt_template
        are given, they override the stored template's own text for this
        render only (e.g. a live preview while editing); the field left
        unset falls back to the stored template's text.
      operationId: renderTemplate
      tags: [templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
                kind:
                  type: string
                  enum: [generic, grafana, github]
                title_template:
                  type: string
                prompt_template:
                  type: string
              required: [payload]
      responses:
        "200":
          description: The rendered title, prompt and any template-execution errors
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                  prompt:
                    type: string
                  errors:
                    type: array
                    items:
                      type: string
                required: [title, prompt, errors]
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: styr_session
    bearerToken:
      type: http
      scheme: bearer
      description: >
        A personal API token (POST /me/api-tokens), sent as `Authorization:
        Bearer styr_pat_...`. Equivalent to sessionCookie for every route,
        except that it is exempt from the X-Requested-With CSRF header
        rule (internal/api/middleware.go's csrfGuard).
  responses:
...
    Forbidden:
      description: The current user may not perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: Not found (or not visible to the current user)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    TemplateInput:
      description: The create/update request body for POST/PATCH /templates(/{id}).
      type: object
      properties:
        name:
          type: string
        workspace_id:
          type: string
        profile_id:
          type: string
        title_template:
          type: string
        prompt_template:
          type: string
        system_prompt:
          type: string
        report_schema:
          type: string
          description: JSON schema text; empty means no structured report is expected.
        shared:
          type: boolean
          description: Owned by nobody rather than the caller. Admin only.
      required: [name, workspace_id, profile_id, prompt_template]
    Template:
      type: object
      properties:
        id:
          type: string
        owner_id:
          type: [string, "null"]
        name:
          type: string
        workspace_id:
          type: string
        profile_id:
          type: string
        title_template:
          type: string
        prompt_template:
          type: string
        system_prompt:
          type: string
        report_schema:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - owner_id
        - name
        - workspace_id
        - profile_id
        - title_template
        - prompt_template
        - system_prompt
        - report_schema
        - created_at
        - updated_at
...